Use "https" scheme for registry_api_url setting in global_config.vocabs.sample
The registry_api_url setting in global_config.vocabs.sample needs to
begin with "https", so that when the API is accessed from JavaScript
code running in the browser, the browser will go ahead and access it
(rather than refusing to access "mixed content", if the portal is
otherwise being accessed using the "http" scheme).

So, change the sample setting from "http" to "https".