Start work fixing date picker after upgrade to AngularJS 1.6
Oh dear, the date picker is broken after the upgrade to AngularJS 1.6.

As a first step to fixing it, stop using the customised template file
that we made for angular-bootstrap 0.10.0.

This broke the modal dialogs. So make them work again.

Next step will be to do customisations of the current version of
angular-bootstrap.