Hide browse tree section if no browse tree to show
After connecting the getting of the browse tree data to the Registry
API, a blank browse tree section was shown in the case that there was
no such data.

Update the visualise directive to examine the new format of the return
value in the case where there is no browse tree to display (i.e.,
because there is no concept tree for the current version).