Link vocabs in search results to view page by ID, not slug
In the search results page, link each vocab to the view page by ID,
not slug. I.e., use the "viewById" controller method.

We may revisit the generated links later. But (once we get Solr
indexing, and Solr searching via the API working) at least we'll be
able to use the results on the search results page!