Remove controller's old add() and disable filter()
I got myself confused about the presence of both add() and addnew().
Since addnew() is basically done, and it almost matches add(), just
remove the existing add() and rename addnew() to add().

Don't delete filter() for now, but "disable" it by having it throw an
exception if it is ever called.