Fix no-longer-relevant redirect to vocabs/editnew/ on login
As I was developing new versions of controller methods, I made an
editnew() method based on edit(), and had logins redirect to that
method.

I subsequently removed the old edit() and renamed editnew() to edit(),
but omitted to change the redirect so that it pointed to edit() again.

Fixed now.