CMS: have TinyMCE patch up relative URLs in href
If you insert literal HTML into the TinyMCE editor that contains an
href attribute whose value is a relative URL, this would be rejected
by the registry's validation using Jsoup.

We might be able to do something about this in the registry, but for
now, have TinyMCE rewrite the relative URL into an absolute one.