Search: restore search filters to URL to cope with change to AngularJS
The upgrade to AngularJS changed the behaviour of the getter
$location.search(). It now returns a _copy_ of the search terms.

(IMNSHO the existing code wrongly took advantage of the fact that it
used to return a _reference_.)

So now, when you change the search terms via $scope.filters, you have
to invoke the setter $location.search($scope.filters).

Do that now.