CMS: when populating REs for project from PP, add all existing details
When populating the list of REs having fetched metadata from PP, if
there's a match for an existing RE, populate it with _all_ the details
of the existing RE, i.e., including phone, email, URLs, and identifiers.
For that, need to go back to the API to get the identifiers.