CMS: Distinguish Exit/Exit Without Saving buttons/behaviours
If the user has not actually changed anything, offer an "Exit" button
that goes straight back to My Vocabs.

If the user _has_ changed something, make the button say "Exit Without
Saving", and require confirmation.

Err on the conservative side of determining whether the user has
"changed" something. For now, that means: using $dirty on the
top-level form, and considering a change to include clicking the
"Save" button on any of the modals.