CC-2217: CMS: Don't try to add new REs for PP if owner not yet known
Situation: from My Vocabs, click "Add a new vocabulary from
PoolParty". Select a project and click "Use this PoolParty
Project". The Owner dropdown appears, with no owner yet selected.  So
far so good.

But: if the metadata fetched from PP contains a
publisher/contributor/creator that the registry doesn't already know
about (i.e., there is no current RE (related entity) with a match
against the title), you get (after a delay of a couple of seconds) an
alert saying there was a failure to create a related entity.

Observing the network traffic shows that the cause is that the
pre-population of the form from the PP metadata happens immediately,
i.e., even though you haven't yet made a selection from the Owner
dropdown. So the API call to create the RE is a request that contains
a blank owner, and the API returns a constraint violation error.

If all the PP project's REs already exist in the Registry, you don't
have this problem.

To fix this:

Now, when the user has made a PP project selection, don't fetch the
metadata straight away. Store the project selection into the scope for
future use. Display the Owner dropdown. When the user confirms the
selection, _then_ fetch the metadata from PoolParty.