CMS: on deprecate action, redirect to view page
After clicking "Publish" on the CMS page, you are sent to the view
page.  But if you clicked "Deprecate", you were left on the CMS page.
Now instead, go to the view page in this case as well.

This "works around" a defect in the CMS, namely, that after a "save"
action, the CMS page is not repopulated with the updated data. Hmm, so
this also still applies for saving to draft.