CC-2162: Use credentials to add subscriptions
The Registry now requires credentials when adding a subscription.

So now, provide them.

Instead of the browser's AngularJS controller invoking the API methods
to add subscriptions, have that done by the PHP controller.

Define a "default" username/password for the Registry in
global_config.php to use, if the portal user is not already logged in.