Fix composer dependencies so they work on production
During deployment to production, composer install failed as follows:

(After parking the old composer.lock)
composer install
Loading composer repositories with package information
Installing dependencies (including require-dev) from lock file
Your requirements could not be resolved to an installable set of packages.

  Problem 1
    - Installation request for phpdocumentor/reflection-docblock 3.3.2 -> satisfiable by phpdocumentor/reflection-docblock[3.3.2].
    - phpdocumentor/reflection-docblock 3.3.2 requires php ^5.6 || ^7.0 -> your PHP version (5.5.38) does not satisfy that requirement.
  Problem 2
    - phpdocumentor/reflection-docblock 3.3.2 requires php ^5.6 || ^7.0 -> your PHP version (5.5.38) does not satisfy that requirement.
    - phpspec/prophecy 1.7.6 requires phpdocumentor/reflection-docblock ^2.0|^3.0.2|^4.0 -> satisfiable by phpdocumentor/reflection-docblock[3.3.2].
    - Installation request for phpspec/prophecy 1.7.6 -> satisfiable by phpspec/prophecy[1.7.6].

Sigh, we only have PHP 5.5 on production.

Had a look in the old composer.lock to see how
phpdocumentor/reflection-docblock and phpspec/prophecy come in to
play. It seems they are dependencies of phpunit.

So in composer.json, in the require-dev section, put in the specific
versions of prophecy and phpunit as were successful (i.e., as in the
old composer.lock).