Major cleanup, removing many RDA-only files
Do a major cleanup of the repository, removing very many RDA-only
files.

Some patches were needed to make this work:

* Tidy composer.json
* Don't use illuminate library methods in vocabs controller
* Tidy footer.php
* Remove "eloquent" methods and their invocations
* Remove 'registry' from list of modules in global_config.php

Also take the opportunity to remove compass as a dependency in the
top-level Gruntfile.js; we can just rename the one .scss file to
a .less file, and that works just fine.