CC-2313 Cope with missing server vars REQUEST_SCHEME, HTTP_HOST
CC-2313
CC-2373

When computing the "HTTP referrer" value to pass to Registry requests
(in the "portal-referrer" header), cope in the cases that either of
the HTTP server variables REQUEST_SCHEME and HTTP_HOST is missing.

If REQUEST_SCHEME is missing, see if HTTPS is present. If that is also
missing, fall back to "http".

If HTTP_HOST is missing, see if SERVER_NAME is present. If that is
also missing, fall back to "localhost". (But in that case, you really
should fix the HTTP Server configuration by setting ServerName.)