Vocabs Widget: return correct Content-Type; test PHP version correctly
Fix a couple of long-standing defects in the Vocabs Widget proxy code:

1. The proxy returned Content-Type "application/json". But this is
wrong: the body is not valid JSON, but JSONP, which is really
JavaScript. So return Content-Type "application/javascript" instead.

2. The display() function had this test:
   $this->jsonData = (defined(PHP_VERSION_ID) && PHP_VERSION_ID >= 50400) ?

But this is wrong: the defined() method expects a string parameter.
In fact, the test always evaluated to false, even though the
PHP_VERSION_ID constant _was_ defined, and had a value > 50400 !
Pass a string parameter instead:
   $this->jsonData = (defined('PHP_VERSION_ID') && PHP_VERSION_ID >= 50400) ?