CC-2442: don't escape description in VRV preview popup
On a vocabulary view page, related internal vocabularies are shown in
the right-hand-side sidebar.

If you click on one, you get a popup.

The popup includes the vocabulary description.

The content of the description has HTML tags escaped, so that you see,
e.g., "<b>bold</b>".

This is a leftover from pre-registry times, where the content could be
"anything", i.e., either plain text or HTML.

Now that it is guaranteed to be a valid HTML fragment, don't do
escaping of the description in these popups.

Having removed the escaping, there was an unexpected and undesired
side-effect: the display of the popups changed, so that the text would
often wrap "too much", making the width of the popup unnecessarily
small, rather than expanding to use the full value of the max-width
setting.

It seems this is a feature/defect of qTip2, and (I checked) is not
changed in the latest release (3.0.3).

So, after some trial and error, I found that changing the position
adjustment method from 'shift' to 'flip' in the horizontal direction
seems to have done "enough" to make the popups usable again.