Have controller page() method default $slug to empty string
The URL .../vocabs/page/ invokes the controller method page($slug),
but $slug is then undefined, and you get all sorts of errors in the
logs, and instead of our nicely-formatted soft 404 page, you get a
sparse internal Blade error.

Change the definition of the method to begin:
    public function page($slug = '')

so that $slug defaults to an empty string, instead of being
undefined. Then the method proceeds in a "normal" way, through to
monolog-logging the soft 404 and returning the correct soft 404 page.