CC-2456 have grunt copy .min.js.map files so they can be found
When viewing a page (e.g., in Safari) with the developer tools open,
the browser loads .min.js.map files.

But some of those files can't be found, when you're running in
production mode and the sourceMappingURL directives being followed are
in the consolidated lib.js.

So have grunt copy the .min.js.map files in question to the same
directory as lib.js, where they will be found.