CC-1488 Add more vocab data to log entries for related entity/vocab
When adding an analytics log entry for the preview of a related entity
or vocabulary, include not only the vocabulary id, but also status,
title, owner, and slug.

Remove the use of "v_id" attributes in the "a" elements, as we now
have the "vocab:id" meta field in the page head.