CC-1488 Don't log portal_cms action for page load for add
In the previous commit, I commented out the analytics log entry for
the page load of the CMS for editing an existing vocabulary, since it
is (for now) the confirmation of the edit that is to be logged.

I should also have commented out the analytics log entry for the page
load of the CMS for adding a new vocabulary. Do that now.