Require user to press Enter/click Search to do a search
Remove the auto-search feature, where the user just had to stop typing
their search term, for the search to kick in.

Now, the user must either press Enter or click the Search button to
start a (new) search.

This addresses the problem that the Registry search method was often
invoked twice for a search, which affected analytics logging.