CC-1488 Harmonize user component of analytics log entry with Registry
The user component of analytics log entries had "userid" and
"username" fields. That's already confusing, and what was even more
confusing was that the latter field was used to store the user's full
name.

So use more conventional field names: "username" and "fullname". Store
the user's username in "username", and their full name in "fullname".

The use of "username" now matches the Registry's behaviour.