Define onclickURL() for use by all versions
Sigh, I defined onclickURL() within the condition for the existence of
a current version, but I also use it for all other versions. So if
there is no current version, onclickURL() doesn't get defined, and the
view page breaks.

So, move up the definition of onclickURL() so that it is always
defined.

I also moved up getIdForSissvocAccessPoint(), even though that is
(currently) only needed for the current version.