CC-1845 Re-implement concept browse tree visualisation
Re-implement the concept browse tree visualisation using fancytree.

We mimic the look-and-feel of the original visualisation, including
the CSS that shows lines linking parents with children.

For browse data that includes support for notations, there is a
dropdown to switch the sort field, and display of notations within
labels can be enabled/disabled using a checkbox.

Filtering expands as necessary to show matches; matches against the
node label are highlighted.

Display of polyhierarchies is supported; nodes that exist in more than
one place include a link in their tooltip to filter to show all
instances of that node.

Note: all this relies on the corresponding changes to the Registry!