CC-1845 Adjust wrapping of long labels in browse visualisation
In the new browse visualisation, the wrapping of long labels was done
in a way so as to make it straightforward to observe the wrapping
behaviour during testing. The result was that the wrapping happened
way too soon, i.e., without using all the available width of the
panel.

Adjust the wrapping so that it only kicks in when necessary.