CC-2524 CC-2525 Fix widget expand/collapse behaviour
The vocabulary widget's tree mode is broken. If you expand/collapse a
child at a lower level, then collapse/expand at the top level, the
lower level is shown expanded, but with the "+" icon next to it.

To reproduce:

1. Go to the Widget Explorer.
2. Select the vocabulary "Australian Land Use and Management
   Classification".
3. Click in the "Sample Input Element" text area to open the tree.
4. Click the first "+" icon next to "Conservation ..." to open its
   children.
5. Click the "+" icon next to the first child "Other minimal use" to
   open its children. That icon now changes to "-".
6. Click that second-level "-" icon to close those children.
7. Click the top-level "-" icon to close those children.
8. Now click the first "+" icon next to "Conservation ..." to open its
   children.

Observe that the children of "Other minimal use" are also shown! But
note also that the icon next to "Other minimal use" is still a
"+". (If you click it, it changes to "-".)

The correct behaviour of step 8 is to open the children of
"Conservation ..." but to show no lower-level children.

This is caused by the use of the jQuery method find(), where it should
be children(). (Using find() gives all descendants that match the
filter; but in this case only direct descendants (children) should be
matched.)

So, replace the two uses of find() with children().