CC-1845 Do HTML escaping of content in browse and its tooltips
Sigh, hadn't yet applied HTML escaping to vocabulary content in the
browse visualisation and its tooltips. Do the escaping, where needed.

For further investigation: whether this is adequate. To be specific:
there are some remaining uses or IRIs that are not escaped, but which
are injected into JS strings used in attributes.