CC-1845 Do more HTML escaping of IRIs in browse tooltips
Further to the comment in the earlier commit:

"For further investigation: whether this is adequate. To be specific:
there are some remaining uses of IRIs that are not escaped, but which
are injected into JS strings used in attributes."

Indeed, do more escaping of IRIs within tooltip content. Note: if an
IRI breaks the syntax rules, it may get through into the Sesame repo,
but it may be impossible to search for it or to bring up the SISSVoc
concept endpoint!

Example of something that breaks in this way:
<https://editor.vocabs.ands.org.au/ANDSRichardtestHTMLinlabels/3&amp;entity"quote>