CC-1845 Use e.currentTarget instead of e.target in two places
If you use use e.target in an event handler, you can get something
other than the element you are defining the handler for! It can be a
nested element.

Use e.currentTarget instead of e.target in two places, to make sure we
get the DOM element we want.