CC-2571 Define AngularJS filter to help with string/array processing
The new "current search" box will display the list of active filters.
For each type of filter, the value can be either a string or an array.

We want to iterate over filter values in such a way that we don't want
to code specially for the two different types of value; in particular,
we want to treat a string value as though it were an array containing
one element.

So, define an AngularJS filter that converts a string value into an
array containing one value, which is that string.