CC-2571 CC-2572 Add "Current search" box and support clearing
CC-2571:

Add a new "Current search" box on the search results page.

In the "Current search" box, show the active search query and all the
active filters.

Show all the facet information in a "Refine search results" box.

In the "Refine search results" box, use checkboxes at the left-hand
side, instead of an "X" after the facet value.

Show the facet counts at the right-hand-side.

Use (almost) the full width of the screen for the search results page.

CC-2572:

In the "Current search" box, include a "Clear search" button to clear
everything.