Revert "CMS. Version release date display now formats to yyyy-mm-dd correctly upon saving and reopening"
This reverts commit aa519d8dff1c9d32c7d3526fb11267c6ad9f8a77.

The commit aa519d8dff1c9d32c7d3526fb11267c6ad9f8a77 was done in
response to a Trello ticket:

"CMS-Version : Upon reopening a version for editing the date has
changed from yyyy-mm-dd to Mon Jan 29 2018 00:00:00 GMT+1100 (AEDT)"

It is not clear why that happened.

However, the code change in the commit is wrong: it rewrites a value
"2019" or "2019-01" from the database into "2019-01-01". It precisely
should _not_ do that.

The code in the commit has been completely reverted.

After reversion, the behaviour reported in the Trello ticket does not
reoccur.