Truncate portal-referrer header so it won't exceed HTTP header limit
We send portal referrer information as an HTTP header.
There is a limit on the length of each HTTP header.

It's possible to exceed this limit. This can happen if you add a
search filter that has a long value.

So, now _always_ truncate the portal referrer header so that it will
always be less than (say) 4096 characters.