CC-2599 Prevent jump to browse filter text input
On page load of the view page, if there's a browse visualization
present, but there's so much vocabulary metadata that the
visualization isn't visible, the browser jumps (scrolls) to make the
text input for filtering visible.

Prevent the jump by having the visualization load with the text input
hidden, then only enable at the end of the controller's
initialization.