CC-2600 Inhibit restore scroll location going back search results
You can use the search results page to step through results pages,
change the sort order, add filters, etc.  When you do any of that, the
browser "remembers" where you have scrolled to on that page of search
results, and by default, if you use the browser's back button to come
back, the browser restores your scroll position. We already have code
in place that immediately animates a scroll to the top of the page,
but this looks unattractive.  Prevent the browser's initial restore of
the scroll position in this case, so that on browser back, the page of
results is shown scrolled to the top.

This is currently done using a browser-dependent feature "manual
scroll restoration". It is not (yet) supported on IE or Edge.