CC-2619 CC-2620 Facet filtering: show unused values, "+x" counts
CC-2619: We now do facet filtering as a conjunction of
disjunctions. So we now need to offer facet values that _aren't_ used
in current search results. The Registry search API method now includes
the data we need, so use it.

CC-2620: For filter values that aren't included in current search
results, or that are but which aren't current filter selections, offer
a "+x" value next to the checkbox to indicate the effect on the total
number of search results if this filter were to be activated.