CC-2600 Allow scroll restoration when coming back from a view page
Further to commit 67c0ca1b516c3774e0d9b220dfb131a43683eb25, now _do_
allow restoration of the scroll position when coming back (i.e., using
the browser's back button) to a search results page from a view page.

Due to the inability to distinguish forward and back events, the
scroll position is also restored when going forwards through search
results pages to the _last_ in the contiguous series of search result
pages.

As before, this is currently done using a browser-dependent feature
"manual scroll restoration". It is not (yet) supported on IE or Edge.