CC-2620 Don't clear facet data before new search
As part of commit bde19d13c0495d7ebe7f83f948029eab6d054e0f, I added a
line to avoid "blinking" of some warning icons in the breadbox.

I faild to notice that that line has the unwanted side-effect of
_causing_ the entire "Refine search results" to blink instead!

Since we decided not to proceed with the warning icons, we may as well
comment out that line.

CC-2619