CC-2612 Resource search: don't offer reset for default collapse/expand
If you reset search, then clicked the "Search" button, the "Reset
search" button became active. This was because the anyFilters() method
hadn't been updated to take the r_collapse_expand filter (for the
collapse/expand setting) into account.

Update anyFilters() to ignore the r_collapse_expand filter when it has
its default value, so that the "Reset search" button isn't active just
because the collapse/expand setting is the default value.