CC-2612 Resource search: invoke callback from bottom nav tabs
If you changed search result tabs by clicking one of the tabs in in
the top navbar, the tabChanged() method was invoked.  But the
tabChanged() method _wasn't_ invoked if you clicked one of the tabs in
the _bottom_ navbar, which meant that the activeTab filter didn't get
updated properly, and the page could get "confused".

Now invoke the tabChanged() callback from the bottom nav tabs as well.