Search results: don't try to scroll to top twice
On the search results page, it seems it's convenient to call
scrollToTop() in multiple places to "make sure" that we do that
scrolling.

But this can mean that scrollToTop() is invoked twice, and if you
start using the mousewheel during scrolling, you have to "fight" the
browser.

So, implement a primitive (but good enough) lock so that we don't try
to do the scrolling twice.

(Yes, there's now a race condition, but let's see if it ever arises in
practice.)