CC-2612 Resource search: for other vocabs, show correct version title
Due to a strange/stupid copy/paste error (?), if you expanded a "View
other vocabularies" section, and there were multiple versions for a
vocabulary, each version in the list was shown with the version title
of the _first_ version in the list for the vocabulary, not with its
own version title.

Display the correct version title for each version.