CC-2673 Allow widget proxy to follow redirects
A config change to some servers broke the widget. On some servers,
there is now a redirect from http -> https for all requests.

The widget proxy is making requests of URLs with scheme http, but it
doesn't expect to get back responses with status 301.

So now, allow the proxy to follow up to 5 redirects ... which should
be plenty.