Search: use count_only filter when possible
The Portal requests two searches for every change in filters: a
registry and a resource search. But because of the Portal's tabbed
interface, the Portal user can only see one set of results at a
time. The Registry now supports a "count_only" filter that only gives
the numFound value, but no search results (or facet counts or
highlighting, etc.). So use that filter whenever possible.