Analytics: fix generation of logAccessPoint URLs for view page
Oops, the generation of onclick URLs for Sesame downloads was broken.
The definition of the onclickURL function and its use within the blade
relied on the last parameter being the access point URL; for a Sesame
download, the file type was appended to the return value of the
onclickURL function.

Now that there's a query parameter after the access point URL, the
file type was getting appended to the new last query parameter, rather
than to the access point URL!

Do this "properly": pass in the file type as a parameter.