CC-2699 Concept search results: remove bogus "hiddenLabel" text
RVA concept search results include values for SKOS prefLabels,
altLabels, and hiddenLabels.

In the case of hiddenLabels it seems I left a trace of the first draft
of the implementation, because there is a bogus additional bit of text
"hiddenLabel:" shown.

So, e.g., "SKOS hiddenLabel: hiddenLabel: 5" instead of just "SKOS
hiddenLabel: 5".

Remove the extraneous bit of text from search-view.blade.php.

RVA-38