Don't set response headers to JSON for authentication/social
The authenticate() method always sets reponse headers for JSON, even
for social logins. But social logins send a redirect to a different
URL, and the response for that URL is in HTML. A browser _may_ get
confused by that. (I've experienced this today, in one profile in
Firefox.)

So don't set those response headers when redirecting to a provider
for a social login.