CC-2734 Restore fuzzy matching by email for Rapid Connect logins
This is a hotfix for Release 37.1.

The refactor of RDA's authentication subsystem in September 2018
removed some fuzzy matching of Rapid Connect logins. It used to be the
case that if the role ID (which is the SHA1 of the user's
"edupersontargetedid" value) changes – as it does, for example, when
the registered callback URL changes – there is fallback to matching by
email address, and then by display name.

But no longer. Now, only exact matching on role_id applies.

RDA's code was migrated into RVA in January 2020, and it is now also
affected.

We have now changed the Rapid Connect callback URL for RVA production,
and so the Rapid Connect authenticator no longer finds existing users
in the roles database, so a new user gets created, i.e., with a
different role_id, but which of course has none of the functional and
organisational roles of the old user role.

The new code's role creation also has the defect that it does not
store the email address that comes from the user's AAF profile. And
indeed, we find in the roles database that recent (i.e., since
September 2018) logins have a blank email value.

I also noticed in the code that when looking for an existing role,
there are some database queries that don't check the value of the
enabled column; these are defects.

We're in luck to patch up the database: the column oauth_data is JSON
data that includes the email address. Therefore, we are able to patch
up the roles data for existing user roles for which the value of the
email column is blank.

Do the following:

* Patch the code to restore fuzzy matching, at least by email address.
* Store the email address in the email column, when creating a new
  role.
* When querying, add the test 'enabled' => DB_TRUE to the where clause
  of database queries for existing roles.
* Still need to do this: in all our roles databases, do a cleanup of
  the roles data, making sure that there is only one enabled role for
  each user, and that it has the correct value in the email column.