CC-2717 CMS: fix hover tooltips on PP project titles
Using My Vocabs, select "Add a new vocabulary from PoolParty". You
then enter something in the search box, and you get a dropdown with a
list of matches.

Now hover over one of the entries in the list. You get a tooltip
containing the text:

{{match.label}}

The content of the tooltip is supposed to be the project name.

The project selection interface is an instance of the "uib-typeahead"
widget, and the template for that widget uses the default escape
mechanism for AngularJS, which is the double-brace syntax.

But in the RVA code we have overridden that syntax, and we use
double-brackets instead.

So override the template and use double-brackets instead.

RVA-88