CC-2522 Browse: support concept schemes, collections, URLs
Support "format version 3" of the ConceptTree version artefact.
Support:
* Display of concept schemes
* Display of collections
* Hyperlinking IRIs are valid URLs

Show references using the fa-clone icon in the node title and in the
tooltip.

Add CMS version browse flags to support enabling each of the above.

Continue to support the display of version artefacts that are in
"format version 2".

The visualization of format version 3 does not suffer from the problem
described in CC-2760; The visualization of format version 2 still does.

CC-2523
RVA-43
RVA-44
RVA-45
RVA-46
RVA-47
RVA-48