CC-2760 Expand concept_refs linked by concept_refs
The browse tree display for format version 2 doesn't always display
all of the nodes it should, where a node and one of its children is
involved in multiple polyhierarchies.

The format version 2 code doesn't unpack references (a.k.a. "clones")
recursively. As nodes are cloned and added to the tree, any child
nodes that are themselves references must also be unpacked.

(We already fixed this for format version 3.)

RVA-104