CC-2522 Don't lose highlighting of children when not treefocus
If you open nodes of the browse hierarchy, so that a node and its
children have the "active" highlighting, then click in the filter text
area, the (top-most) active node retains its highlighting, but the
children lose it.

This is because when clicking away, the active node loses the
.fancytree-treefocus CSS class.

Change the CSS rules for the child nodes so that they don't require
.fancytree-treefocus.

Remove the type-specific rules for the "active" nodes that are already
covered by a more generic rule.