CC-2733 Have authenticator redirect to My Vocabs
The authenticator's 'redirect_to' is meant to be the user's "home
page".  This value comes into play when, for example, the user has a
browser tab on their "home page", but leaves it long enough for their
authentication cookie to expire, and they then reload that page and
log in again.

In practice, the existing value (base_url()) meant that you ended up
on the legacy "Dashboard" page.  Set the value so that you end up at
My Vocabs instead.

RVA-103