CC-2756 If reopening version modal after save, restore release date
The CMS uses a datepicker widget in two places: one at the top level,
for the vocabulary's creation date, the other on the version modal,
for the version's release date.

The datepicker widget is implemented using a third-party library, and
we have to work around some of that library's behaviour in order to
support the various date formats we want, i.e., YYYY, YYYY-MM, and
YYYY-MM-DD.

Because the widget is in some ways more generic than we require/use,
it maintains a representation of a full date and time using a
JavaScript Date instance, which represents a date/time to millisecond
precision.

If not configured otherwise, the widget shows a human-readable version
of this representation in its text area, e.g., "Thu Aug 13 2020
00:00:00 GMT+1000 (Australian Eastern Standard Time)". Our own code
"hides" this internal representation and presents to the user the
value they have entered in the supported format, e.g., in this case,
"2020-08-13". This "hiding" of the widget's internal representation is
also important for communicating with the Registry, which expects the
date values also to use only the supported formats.

Historically, we've had a lot of pain and exerted a lot of effort to
make sure that these fields work reliably, but it seems there are
still cases where the internal representation turns up where it's not
wanted.

Here is a scenario I encountered:

1. Edit an existing vocabulary that has an existing version.
2. Click the version title to open the version modal for the version.
3. Observe that the release date text area shows the value correctly.
4. Click the version modal's "Save" button.
5. Now click the same version's title to open the version modal again.
6. Observe that the release date text area now shows the internal
   format.
7. Click the version modal's "Save" button.
8. Scroll down to show the publication buttons, and click the Publish
   button.
9. You get a validation error: "Version release date must be specified
   and be a valid date; value must be either 4, 7, or 10 characters
   long". You get this error despite the fact that the release date was
   valid, and you didn't do anything to change it.

The same phenomenon can be observed with a new vocabulary and/or a new
version; the steps given above are just what seem to be most
convenient to demonstrate the problem.

What's going on?

If you create a version, or edit an existing version, the version
modal's code sets a special field of the version's internal
representation that serves as a "hint" to the top-level save action
(i.e., invoked by the Publish button) that it must throw away the date
value provided by the widget's internal representation, and "revert"
to the date value stored in the version modal's text field. Under
"normal" circumstances, i.e., if you've only opened a version once,
the text field will still have the correct value by the time you click
the Publish button.

But if you open the version a second time, that "patch up" code hasn't
been executed yet, and the version's copy of the release date is still
the widget's internal value, and that's what is then copied into the
text field.

The solution: since the version modal can tell, when it is loaded, if
the version has been "saved" previously in this way, add some release
date "patch up" code to the initialization of the version modal dialog
for this situation.

RVA-105