CC-2627 Restore draggability to CMS modal dialogs
Sigh, after the rebranding, the CMS modal dialogs are no longer
draggable.

Commit dc484cb59121222b82e43870988145261e7d1b39 moved the loading
of jquery-ui.min.js to just after jquery.js.

It seems this means that the JQuery UI methods get "lost". Why?
Perhaps when AngularJS is loaded, but I'm not sure.

In any case, it's not obvious where to move the loading of
jquery-ui.min.js to make everything work again.

So instead, take a copy of the value of $ just after
jquery-ui-min.js is loaded, and use that in the code instead.

The copy is called jQueryWithUI. So where JQuery UI is needed,
say jQueryWithUI(...) instead of $(...).

RVA-40