CC-2522 Browse: if mayResolveResources present, don't assume equals true
Oh dear. Never write JavaScript.

The test of the mayResolveResources flag was consistent with other
tests, but still wrong. In the case of this particular flag, the
Registry may provide the flag, but set to false.

But the controller assumed that _if_ provided, it would be set to
true.

Remove that assumption, and test the value properly.

RVA-48
RVA-149