CC-2760 Browse: ensure clone's descendants won't get cloned
When we clone a node, we copy the children of the canonical instance
of the node. We already took care of the case that some of that node's
children may also be canonical, i.e., having a type _not_ ending in
"...ref". But we didn't take care of the case that some of that node's
_granchildren_ or more deeply nested descendants may also be
canonical.

So when we create the new node, add the sentinel property
isClonedFromRef not only to that node, but to all the new node's
descendants as well.