CC-1524 New implementation of language selection and display
Add a new implementation of language metadata selection and display.

In the CMS, use a ui-select for language selection. Support entering a
BCP 47 tag, which will be looked up in the Registry.

On the view page, show the BCP 47 tag for each language.

Ensure pre-filling the CMS page from PoolParty works for all of
PoolParty's supported languages.

RVA-41