CC-2822: Fix canonical search URL /vocabs/search
Although the URL /search works, I broke the "canonical" path to
the search page, i.e., /vocabs/search, in July 2019, as part of
commit 422b67e91267f8220ad0f18d8f2dc9b8c3856040.

As part of that commit, I added some JS code in two places in the
search controller that answered the question "are we on the
search results page?". But in both cases, the code to ask that
question was wrong.

The code said (in two places):
    if (! $location.absUrl().slice(base_url.length).
        startsWith('search')) {

Oh, and uh-oh.

Yes, indeed, the search page is accessible as /search, but it's
also supposed to be accessible as /vocabs/search. It's the latter
path that the link on one of our public-facing documentation
pages uses.

(The extra path component /vocabs is the name of our (PHP)
controller, so /vocabs/search means the search() method of the
vocabs controller. But the "legacy" RDA
dispatcher (applications/portal/core/controllers/dispatcher.php)
routes the path /search to the search method of the "default
model", which is defined in global_config.php to be -- you
guessed it -- "vocabs". It seems we have been using
this "shortcut" ourselves (i.e., for all RVA's own links to
itself), which is why I/we didn't notice this problem until now.)

Make the broken URL work again: fix the code to support both
/vocabs/search and /search.

RVA-211