CC-2906 Force line breaks when needed
In the view page and in the CMS, there are some fields where we need
to force a line break if the content just won't fit.

Use a combination of class="break" and setting max-width in order to
force line breaks in those fields.

Fields affected:

View page:

* Acronym, Description, Revision Cycle, Notes, Top Concepts, Version
  title

CMS:

* Versions TITLE

RVA-243