CC-1294 CC-2901 Specify set_cookie expiry correctly
While working on CC-2901, I have discovered that I made a mistake when
I applied the fix needed for CC-1294.

It turns out that I overlooked a crucial difference in the behaviour
of PHP's built-in setcookie() and the CodeIgniter wrapper
set_cookie(): the two functions interpret the value of the expiry
parameter in a very different way.

The PHP built-in function setcookie() function's expiry parameter is
"The time the cookie expires. This is a Unix timestamp so is in number
of seconds since the epoch."
(https://www.php.net/manual/en/function.setcookie.php)

But the CodeIgniter set_cookie() function's expiry parameter is the
"Number of seconds until
expiration". (https://codeigniter.com/userguide3/helpers/cookie_helper.html)

Oops.

So when changing setcookie to set_cookie, I should also have changed
"time()+3600" to just "3600".

This explains why, whenever you actually observe the
ands_auth_redirect cookie, it has an expiry date in the year 2075 ....

Since we delete the ands_auth_redirect cookie on the next response,
this defect is (mostly) harmless.

But fix this now: just specify the number of seconds, without adding
time() to the value.

I'm also changing the number 3600 to 600 in order to reflect that this
cookie is intended to be very short-lived.

RVA-244